feat: Add Invoke-PaketForAl function for easy use of Paket CLI
This commit adds a new PowerShell function called Invoke-PaketForAl. This function provides an easy way to use the Paket CLI for downloading NuGet packages. It accepts various parameters such as the command to run, the project path, and the Paket executable path. It also supports using packages as a cache.